Admin page is functional
CRUD operations for all tables are working, with links to the ones the client
may actually need on the index page.

These links include 'Create Building', 'Create Floor', 'Create Apartment', and
'Create Semester'. Semesters will probably be renamed to 'Contract Period', or
something like it, to reflect that breaks are covered there as well.